+++ tests/archives/cvt_archive_test.py
@@ -101,9 +101,8 @@ def test_add_single_and_overwrite(data):
                                                        arbitrary_metadata)
     assert status == AddStatus.IMPROVE_EXISTING
     assert np.isclose(value, high_objective - data.objective)
-    assert_archive_elite(data.archive_with_elite, arbitrary_sol,
-                         high_objective, data.measures, data.centroid,
-                         arbitrary_metadata)
+    assert_archive_elite(data.archive_with_elite, arbitrary_sol, high_objective,
+                         data.measures, data.centroid, arbitrary_metadata)
 
 
 def test_add_single_without_overwrite(data):